pyproject.toml
[tool.poetry]
name = "masters-calorie-tracking-bot"
version = "0.1.0"
description = "A python hw"
authors = ["Nikolay Chechulin <nchechulin@protonmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
logger = "^1.4"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^23.3.0"}
bandit = "^1.7.5"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pre-commit = "^3.3.2"
pep8-naming = "^0.13.3"
flake8-absolute-import = "^1.0.0.1"
flake8-new-union-types = "^0.4.1"
flake8-annotations-complexity = "^0.0.7"
poetry-plugin-export = "^1.9.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
